Add files via upload
New points:
- main_loop cleaning: in main_loop.py there are only the main loops and the algorithm 
- simulation: to call on the cluster, it runs the simulations and save the results in csv
- data_analysis: where we can writhe our code to make data analysis
- save_files: where the functions that the save files are. So you can change them in order to be fine for the clusters
- plots_functions: where there are the plotting function 
- find_paramters_simulation: to call on the cluster, it runs the simulations with all the parameters combinations and then save everything (parameters, averages results and non spreading values) in csv files
- data_analysis_parameters: where we can writhe our code to make data analysis for the parameters

Basically to use it on local machine, first run a simulation (simulation or find_paramters_simulation) and a data analysis (data_analysis or data_analysis_parameters).
