+++ death_parameters_results.csv
@@ -1,0 +1,5 @@
+,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89
+0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.125,0.0,0.0,0.0,0.0,0.125,0.125,0.0,0.0,0.0,0.0,0.125,0.0,0.0,0.125,0.25,0.125,0.25,0.125,0.25,0.25,0.125,0.375,0.25,0.75,0.0,0.125,0.0,0.125,0.125,0.5,0.375,0.375,0.375,0.125,0.375,0.0,0.125,0.125,0.375,0.5,0.125,0.125,0.375,0.375,0.5,0.125,0.125,0.25,0.0,0.0,0.25,0.125,0.0,0.125,0.25,0.5,0.0,0.25,0.0,0.375,0.25,0.25,0.0,0.125,0.0,0.0,0.0,0.0,0.125,0.0,0.0,0.125,0.0,0.0,0.125,0.25,0.0
+1,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.1,0.0,0.0,0.0,0.0,0.0,0.1,0.0,0.0,0.1,0.2,0.0,0.0,0.1,0.1,0.1,0.0,0.1,0.0,0.2,0.0,0.0,0.1,0.0,0.2,0.1,0.1,0.0,0.2,0.1,0.3,0.2,0.3,0.2,0.1,0.1,0.1,0.2,0.2,0.5,0.3,0.1,0.0,0.1,0.0,0.5,0.3,0.3,0.0,0.0,0.0,0.0,0.3,0.1,0.1,0.0,0.0,0.3,0.1,0.0,0.3,0.0,0.0,0.1,0.0,0.0,0.2,0.0,0.1,0.1,0.1,0.2,0.0,0.2,0.1,0.0,0.1
+2,0.0,0.0,0.0,0.0,0.0,0.14285714285714285,0.0,0.0,0.14285714285714285,0.0,0.0,0.0,0.0,0.0,0.0,0.14285714285714285,0.0,0.14285714285714285,0.0,0.0,0.0,0.42857142857142855,0.0,0.0,0.2857142857142857,0.0,0.14285714285714285,0.14285714285714285,0.0,0.0,0.0,0.0,0.14285714285714285,0.0,0.0,0.0,0.0,0.0,0.0,0.2857142857142857,0.0,0.14285714285714285,0.0,0.0,0.2857142857142857,0.14285714285714285,0.14285714285714285,0.0,0.0,0.14285714285714285,0.0,0.2857142857142857,0.0,0.14285714285714285,0.2857142857142857,0.14285714285714285,0.0,0.14285714285714285,0.0,0.14285714285714285,0.0,0.14285714285714285,0.0,0.0,0.14285714285714285,0.2857142857142857,0.0,0.0,0.0,0.14285714285714285,0.14285714285714285,0.0,0.2857142857142857,0.14285714285714285,0.5714285714285714,0.42857142857142855,0.2857142857142857,0.14285714285714285,0.14285714285714285,0.14285714285714285,0.42857142857142855,0.0,0.0,0.0,0.14285714285714285,0.0,0.0,0.0,0.14285714285714285,0.0
+3,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.125,0.0,0.125,0.125,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.125,0.0,0.0,0.0,0.125,0.0,0.0,0.0,0.125,0.125,0.0,0.0,0.125,0.0,0.0,0.25,0.125,0.0,0.125,0.0,0.125,0.0,0.0,0.0,0.125,0.0,0.125,0.0,0.0,0.125,0.0,0.0,0.125,0.0,0.125,0.125,0.125,0.0,0.125,0.125,0.0,0.125,0.0,0.0,0.125,0.0,0.0,0.0,0.0,0.125,0.125,0.0,0.0,0.0,0.0,0.0,0.125,0.0,0.0,0.0,0.0,0.0,0.125,0.0
