+++ infected_parameters_results.csv
@@ -1,0 +1,5 @@
+,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89
+0,5.0,0.0,0.25,1.5,1.0,1.625,1.125,1.625,3.625,3.125,2.75,4.125,2.125,6.125,4.75,5.875,6.875,9.25,8.5,10.625,12.125,13.125,14.0,17.375,15.125,18.625,21.375,19.25,23.0,22.125,17.625,18.0,19.25,22.125,20.5,21.625,21.375,15.5,19.5,19.625,17.375,18.625,18.25,16.5,15.0,17.125,16.375,17.0,14.875,15.875,11.75,12.0,13.0,12.125,11.125,9.5,10.75,10.75,8.0,7.875,9.0,5.875,7.875,8.5,6.75,7.25,5.625,5.0,7.5,5.125,3.75,6.625,5.625,4.625,4.375,3.875,5.375,3.375,5.25,3.125,3.25,4.25,3.0,2.5,2.75,3.0,2.125,2.625,2.125,1.875
+1,5.0,0.0,0.0,1.1,0.8,0.6,2.0,1.0,2.3,2.3,3.9,2.9,2.9,3.1,3.5,6.0,6.0,4.8,6.2,5.8,6.0,6.4,7.0,8.3,8.2,8.4,8.1,9.5,9.5,8.4,9.7,9.8,9.0,8.6,12.6,9.5,10.0,12.2,10.2,10.5,11.4,10.7,9.5,10.6,10.4,14.0,11.4,10.2,10.8,11.1,11.7,10.5,8.2,8.0,9.3,10.2,8.0,8.9,10.0,10.5,7.8,8.1,9.1,9.0,9.2,9.2,7.5,7.7,8.0,8.3,8.0,6.0,6.3,6.1,6.3,5.9,6.4,4.9,6.1,5.2,5.1,3.9,4.4,4.5,3.7,4.2,4.1,4.3,4.7,5.0
+2,5.0,0.0,0.8571428571428571,0.42857142857142855,1.7142857142857142,1.4285714285714286,3.142857142857143,2.142857142857143,1.0,2.857142857142857,3.5714285714285716,4.0,5.142857142857143,6.857142857142857,5.571428571428571,7.142857142857143,5.857142857142857,6.571428571428571,5.142857142857143,5.285714285714286,4.857142857142857,5.571428571428571,5.0,7.0,5.142857142857143,4.142857142857143,6.571428571428571,7.0,6.428571428571429,9.0,8.142857142857142,10.142857142857142,9.285714285714286,8.857142857142858,6.714285714285714,9.571428571428571,10.428571428571429,10.142857142857142,10.857142857142858,10.0,8.142857142857142,7.714285714285714,10.285714285714286,12.142857142857142,12.142857142857142,9.857142857142858,11.857142857142858,8.714285714285714,11.142857142857142,9.857142857142858,10.714285714285714,9.285714285714286,9.571428571428571,9.285714285714286,10.857142857142858,8.428571428571429,7.285714285714286,9.285714285714286,7.428571428571429,8.857142857142858,8.285714285714286,11.142857142857142,8.142857142857142,9.571428571428571,9.285714285714286,8.142857142857142,8.142857142857142,8.571428571428571,6.571428571428571,8.285714285714286,8.142857142857142,5.285714285714286,4.571428571428571,5.0,7.142857142857143,5.714285714285714,4.857142857142857,4.857142857142857,4.714285714285714,3.857142857142857,4.285714285714286,3.0,3.142857142857143,3.2857142857142856,2.7142857142857144,2.7142857142857144,3.0,2.2857142857142856,2.5714285714285716,2.0
+3,5.0,0.0,0.375,0.125,1.125,0.5,1.75,2.125,1.25,1.875,1.375,2.5,3.875,3.125,4.25,4.125,4.25,4.25,4.25,3.25,4.375,3.875,3.25,3.625,5.25,5.25,5.375,2.875,4.125,5.375,4.25,3.0,3.375,4.75,3.5,4.75,5.25,3.375,5.375,3.5,4.125,5.875,5.125,3.5,4.125,5.875,4.125,2.875,3.25,5.25,4.625,3.25,4.25,3.5,3.625,3.875,5.5,3.625,3.125,5.375,4.125,2.625,3.25,3.375,4.0,3.0,3.0,4.25,4.0,3.375,5.0,4.75,4.25,3.125,3.625,3.5,3.25,2.25,1.375,3.625,2.125,2.75,2.0,2.5,1.5,2.125,1.875,3.125,0.625,3.125
