+++ infected_results.csv
@@ -1,4 +1,6 @@
 ,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89
-0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,8.0,0.0,0.0,0.0,0.0,3.0,4.0,0.0,5.0,3.0,1.0,8.0,4.0,6.0,4.0,5.0,5.0,3.0,4.0,5.0,7.0,6.0,6.0,14.0,7.0,14.0,5.0,14.0,15.0,15.0,16.0,16.0,20.0,24.0,13.0,17.0,31.0,20.0,26.0,19.0,21.0,13.0,18.0,19.0,15.0,14.0,12.0,10.0,12.0,11.0,10.0,6.0,9.0,6.0,7.0,1.0,8.0,7.0,4.0,3.0,7.0,5.0,4.0,4.0,5.0,0.0,3.0,3.0,2.0,1.0,1.0,2.0,0.0,3.0,1.0,0.0,1.0
-1,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,8.0,0.0,0.0,0.0,2.0,2.0,7.0,7.0,4.0,7.0,6.0,5.0,5.0,7.0,8.0,11.0,16.0,22.0,18.0,17.0,21.0,25.0,19.0,37.0,25.0,26.0,31.0,26.0,22.0,24.0,22.0,25.0,17.0,19.0,14.0,10.0,8.0,11.0,10.0,9.0,4.0,7.0,8.0,4.0,8.0,8.0,5.0,3.0,3.0,4.0,6.0,1.0,1.0,2.0,1.0,4.0,5.0,1.0,3.0,2.0,1.0,1.0,2.0,3.0,1.0,0.0,0.0,0.0,1.0,1.0,0.0,1.0,0.0,0.0,0.0,0.0,1.0
-2,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,8.0,0.0,0.0,0.0,1.0,3.0,0.0,7.0,7.0,3.0,6.0,7.0,7.0,7.0,10.0,11.0,15.0,13.0,10.0,19.0,11.0,26.0,22.0,21.0,26.0,19.0,18.0,28.0,21.0,15.0,20.0,23.0,21.0,20.0,23.0,14.0,21.0,19.0,18.0,10.0,14.0,11.0,13.0,7.0,4.0,6.0,9.0,5.0,6.0,8.0,4.0,3.0,5.0,6.0,0.0,1.0,6.0,0.0,2.0,0.0,4.0,2.0,1.0,0.0,3.0,2.0,1.0,1.0,2.0,1.0,3.0,1.0,0.0,0.0,0.0,1.0,0.0
+0,8.0,0.0,0.0,0.0,7.0,4.0,2.0,1.0,8.0,4.0,8.0,9.0,6.0,7.0,18.0,11.0,14.0,16.0,18.0,12.0,24.0,33.0,32.0,43.0,53.0,28.0,26.0,45.0,31.0,27.0,32.0,31.0,31.0,32.0,30.0,21.0,6.0,23.0,37.0,20.0,21.0,23.0,22.0,13.0,13.0,17.0,6.0,6.0,10.0,8.0,2.0,6.0,2.0,1.0,6.0,4.0,7.0,7.0,1.0,3.0,7.0,7.0,3.0,9.0,5.0,1.0,0.0,5.0,1.0,1.0,2.0,2.0,3.0,1.0,3.0,0.0,3.0,1.0,1.0,1.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,2.0,0.0,0.0
+1,8.0,0.0,0.0,0.0,0.0,0.0,2.0,4.0,5.0,3.0,0.0,5.0,3.0,2.0,5.0,3.0,6.0,4.0,1.0,10.0,14.0,12.0,18.0,11.0,12.0,14.0,24.0,34.0,19.0,18.0,29.0,24.0,31.0,23.0,29.0,36.0,34.0,30.0,24.0,26.0,35.0,29.0,29.0,23.0,32.0,27.0,19.0,23.0,19.0,16.0,19.0,15.0,14.0,13.0,10.0,5.0,12.0,10.0,11.0,13.0,11.0,9.0,8.0,4.0,1.0,5.0,4.0,3.0,7.0,3.0,2.0,2.0,2.0,4.0,1.0,3.0,2.0,1.0,2.0,2.0,2.0,1.0,3.0,0.0,1.0,1.0,2.0,1.0,3.0,1.0
+2,8.0,0.0,0.0,0.0,3.0,1.0,1.0,5.0,3.0,3.0,0.0,2.0,1.0,3.0,3.0,2.0,6.0,3.0,3.0,0.0,9.0,8.0,5.0,8.0,8.0,1.0,14.0,20.0,18.0,14.0,27.0,19.0,28.0,26.0,25.0,20.0,29.0,32.0,34.0,35.0,26.0,34.0,22.0,23.0,23.0,27.0,23.0,20.0,21.0,15.0,13.0,15.0,16.0,7.0,8.0,6.0,7.0,4.0,11.0,8.0,5.0,11.0,4.0,9.0,4.0,6.0,9.0,5.0,5.0,7.0,4.0,2.0,3.0,7.0,1.0,1.0,7.0,0.0,1.0,1.0,0.0,1.0,0.0,1.0,0.0,2.0,0.0,2.0,1.0,0.0
+3,8.0,0.0,0.0,0.0,0.0,3.0,0.0,3.0,2.0,0.0,0.0,3.0,1.0,0.0,1.0,3.0,0.0,1.0,8.0,11.0,5.0,8.0,20.0,9.0,14.0,23.0,33.0,15.0,19.0,16.0,24.0,33.0,31.0,24.0,31.0,33.0,23.0,39.0,25.0,27.0,19.0,25.0,24.0,15.0,22.0,19.0,16.0,19.0,24.0,20.0,12.0,10.0,11.0,13.0,6.0,8.0,9.0,3.0,5.0,7.0,2.0,1.0,4.0,4.0,8.0,3.0,5.0,2.0,1.0,4.0,4.0,3.0,2.0,2.0,5.0,2.0,1.0,2.0,3.0,4.0,1.0,1.0,1.0,5.0,1.0,2.0,2.0,2.0,1.0,1.0
+4,8.0,0.0,0.0,4.0,0.0,1.0,7.0,4.0,1.0,3.0,7.0,5.0,6.0,5.0,3.0,5.0,15.0,17.0,7.0,16.0,18.0,21.0,23.0,17.0,24.0,34.0,32.0,27.0,42.0,37.0,26.0,32.0,28.0,44.0,32.0,17.0,27.0,20.0,18.0,14.0,9.0,19.0,12.0,12.0,12.0,11.0,13.0,6.0,2.0,10.0,9.0,6.0,7.0,6.0,5.0,3.0,6.0,2.0,0.0,1.0,0.0,3.0,1.0,0.0,0.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,0.0,0.0,1.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0
