+++ main_loop.py
@@ -9,24 +9,24 @@ from tqdm import tqdm
 from network_generation import *
 from interventions import *
 
-############################## import data ##############################
-
-# create a dataframe for parameter where the index is age
-parameter_age = parameter_importer(pathandfile('parameters_age.csv'))
-# this is to form a new column 'percentage_population' that shows the percentage of the population in certain age range, in case you need it
-parameter_age['percentage_population'] = parameter_age['population_UK']/parameter_age['population_UK'].sum()
-# create a dataframe for parameter where the index is day
-parameter_day = parameter_importer(pathandfile('parameters_day.csv'))
-
-# columns 
-probability_day_death = column_extractor(parameter_day, 'Probability_of_Death')
-probability_day_death = probability_day_death/np.sum(probability_day_death)
-
-probability_day_incubation = column_extractor(parameter_day, 'Incubation_Period')
-probability_day_incubation = probability_day_incubation/np.sum(probability_day_incubation)
-
-
 def simulation(n_days, n_nodes, n_initial_infected, array_network_parameters, array_weights):
+
+	############################## import data ##############################
+
+	# create a dataframe for parameter where the index is age
+	parameter_age = parameter_importer(pathandfile('parameters_age.csv'))
+	# this is to form a new column 'percentage_population' that shows the percentage of the population in certain age range, in case you need it
+	parameter_age['percentage_population'] = parameter_age['population_UK']/parameter_age['population_UK'].sum()
+	# create a dataframe for parameter where the index is day
+	parameter_day = parameter_importer(pathandfile('parameters_day.csv'))
+
+	# columns 
+	probability_day_death = column_extractor(parameter_day, 'Probability_of_Death')
+	probability_day_death = probability_day_death/np.sum(probability_day_death)
+
+	probability_day_incubation = column_extractor(parameter_day, 'Incubation_Period')
+	probability_day_incubation = probability_day_incubation/np.sum(probability_day_incubation)
+
 	
 	network = Network_Generation(n_nodes) # generate the network
 
@@ -217,6 +217,7 @@ def main_algorithm_fast(n_simulations, n_days, n_nodes, n_initial_infected, arra
 
 	for i in tqdm(range(n_simulations)):
 		matrix_infected[i], matrix_death[i], matrix_recovery[i] = simulation(n_days, n_nodes, n_initial_infected, array_network_parameters, array_weights)
+		matrix_infected[i,0] += n_initial_infected	# add initial infections
 
 	return matrix_infected, matrix_death, matrix_recovery, matrix_vaccination
 
@@ -317,188 +318,7 @@ def main_algorithm(n_simulations, n_days, n_nodes, n_initial_infected, array_net
 			network.node_list, matrix_infected[i,j], matrix_death[i,j], matrix_recovery[i,j] = main_loop(network.node_list, total_network)
 			network.node_list, matrix_vaccination[i,j] = vaccination(network.node_list, vaccinations_number_array[j])
 
+		matrix_infected[i,0] += n_initial_infected	
+
 	return matrix_infected, matrix_death, matrix_recovery, matrix_vaccination
-
-def plot_results(name_file, matrix_infected, matrix_death, matrix_recovery, matrix_vaccination, plot_lockdowns = False):
-	"""
-	Plot the simulation results and save the plots
-	Inputs: 
-	matrix_infected      Matrix of infected (n_simulation X n_day)
-	matrix_death         Matrix of death (n_simulation X n_day)
-	matrix_recovery      Matrix of recovery (n_simulation X n_day)
-	"""
-
-	fig = plt.figure(figsize = (14, 8))
-
-	################## Plot day data ##################
-	
-	#### Infections subplots ####
-	ax1 = fig.add_subplot(241)
-	ax1.set_title('Infections')
-
-	# plot each simulation
-	for i in range(n_simulations):
-		ax1.plot(matrix_infected[i], color='grey', linewidth=0.5,alpha = 0.5)
-	ax1.plot(np.mean(matrix_infected, axis = 0), color = 'b') # plot the average
-
-	#### Death subplots ####
-	ax2 = fig.add_subplot(242)
-	ax2.set_title('Deaths')
-
-	# plot each simulation
-	for i in range(n_simulations):
-		ax2.plot(matrix_death[i], color='grey', linewidth=0.5,alpha = 0.5)
-	ax2.plot(np.mean(matrix_death, axis = 0), color = 'r') # plot the average
-
-	#### Infections subplots ####
-	ax3 = fig.add_subplot(243)
-	ax3.set_title('Recoveries')
-
-	# plot each simulation
-	for i in range(n_simulations):
-		ax3.plot(matrix_recovery[i], color='grey', linewidth=0.5,alpha = 0.5)
-	ax3.plot(np.mean(matrix_recovery , axis = 0), color = 'g') # plot the average
-
-	#### plot vaccination ####
-	ax4 = fig.add_subplot(244)
-	ax4.set_title('Vaccinations')
-	for i in range(n_simulations):
-		ax4.plot(matrix_vaccination[i], color='black', linewidth=0.5,alpha = 0.5)
-	ax3.plot(np.mean(matrix_vaccination , axis = 0), color = 'g') # plot the average
-
-	################## Plot cumulative data ##################
-	
-	cumulative_infections = cumulative(matrix_infected)
-	cumulative_deaths = cumulative(matrix_death)
-	comulative_recoveries = cumulative(matrix_recovery)
-	comulative_vaccinations = cumulative(matrix_vaccination)
-
-	current_infected = cumulative_infections -cumulative_deaths - comulative_recoveries
-
-	#### Infections subplots ####
-	ax5 = fig.add_subplot(245)
-	# plot each simulation
-	for i in range(n_simulations):
-		ax5.plot(cumulative_infections[i], color='grey', linewidth=0.5,alpha = 0.5)
-	ax5.plot(np.mean(cumulative_infections , axis = 0), color = 'b') # plot the average
-	ax5.set_title('Comulative Infections')
-
-	#### Death subplots ####
-	ax6 = fig.add_subplot(246)
-
-	# plot each simulation
-	for i in range(n_simulations):
-		ax6.plot(cumulative_deaths[i], color='grey', linewidth=0.5,alpha = 0.5)
-	ax6.plot(np.mean(cumulative_deaths , axis = 0), color = 'r') # plot the average
-	ax6.set_title('Comulative Deaths')
-
-	#### Infections subplots ####
-	ax7 = fig.add_subplot(247)
-
-	# plot each simulation
-	for i in range(n_simulations):
-		ax7.plot(comulative_recoveries[i], color='grey', linewidth=0.5,alpha = 0.5)
-	ax7.plot(np.mean(comulative_recoveries , axis = 0), color = 'g') # plot the average
-	ax7.set_title('Comulative Recoveries')
-
-	#### plot vaccination ####
-	ax8 = fig.add_subplot(248)
-	for i in range(n_simulations):
-		ax8.plot(current_infected[i], color='grey', linewidth=0.5,alpha = 0.5)
-	ax8.plot(np.mean(current_infected, axis = 0), color='black')
-	ax8.set_title('Current Infected')
-
-	################## Plot lockdowns ##################
-	if plot_lockdowns == True:
-		list_subplots = [ax1, ax2, ax3, ax4, ax5, ax6, ax7, ax8]
-		for j in range(8):
-			for i in range(len(days_lockdown_start)):
-				list_subplots[j].axvline(days_lockdown_start[i], color='r', linewidth=0.5,alpha = 0.5)
-				list_subplots[j].axvline(days_lockdown_end[i], color='r', linewidth=0.5,alpha = 0.5)
-
-			for i in range(len(day_school_close)):
-				list_subplots[j].axvline(day_school_close[i], color='m', linewidth=0.5,alpha = 0.5)
-				list_subplots[j].axvline(day_school_open[i], color='m', linewidth=0.5,alpha = 0.5)
-
-	plt.savefig(name_file)
-	return
-
-def save_results(matrix_infected, matrix_death, matrix_recovery, matrix_vaccination):
-	"""
-	Saves the simulation results in csv files 
-	Inputs: 
-	matrix_infected      Matrix of infected (n_simulation X n_day)
-	matrix_death         Matrix of death (n_simulation X n_day)
-	matrix_recovery      Matrix of recovery (n_simulation X n_day)
-	"""
-	# convert to dataframe
-	df_matrix_infected = pd.DataFrame(data=matrix_infected)
-	df_matrix_death = pd.DataFrame(data=matrix_death)
-	df_matrix_recovery = pd.DataFrame(data=matrix_recovery)
-	df_matrix_vaccination = pd.DataFrame(data=matrix_vaccination)
-
-	# save 
-	df_matrix_infected.to_csv("infected_results.csv")
-	df_matrix_death.to_csv("infected_results.csv")
-	df_matrix_recovery.to_csv("infected_results.csv")	
-	#df_matrix_vaccination.to_csv("vaccination_results.csv")
-	
-	return
-
-def cumulative(matrix):
-	
-	cumulative_matrix = np.zeros((matrix.shape[0], matrix.shape[1] + 1))
-	for i in range(n_simulations):
-		for j in range(n_days):
-			cumulative_matrix[i,j + 1] = matrix[i,j] + cumulative_matrix[i,j]
-	return cumulative_matrix
-
-def filter_non_spread(matrix_infected, matrix_death, matrix_recovery, matrix_vaccination):
-	n_non_spread = 0
-	threshold = 0.8*n_days
-
-	for i in range(n_simulations):
-		n_zeros = np.sum(np.where(matrix_infected[i - n_non_spread] == 0, 1, 0))
-		if n_zeros > threshold:
-			matrix_infected = np.delete(matrix_infected, i - n_non_spread, 0)
-			matrix_death = np.delete(matrix_death, i - n_non_spread, 0)
-			matrix_recovery = np.delete(matrix_recovery, i - n_non_spread, 0)
-			matrix_vaccination = np.delete(matrix_vaccination, i - n_non_spread, 0)
-			n_non_spread += 1 
-
-	return matrix_infected, matrix_death, matrix_recovery, matrix_vaccination, n_non_spread
-
-if __name__ == "__main__":
-	n_simulations = 10
-	n_days = 90
-	n_nodes = 5000 
-	n_initial_infected = 8
-	array_network_parameters = np.array([2,2,2,2,2])
-	array_weights = np.array([0.7,0.1,0.1,0.1,0.2,0.2])
-
-
-	# days of lockdown data
-	days_lockdown_start = []
-	days_lockdown_end = []
-	day_school_close = []
-	day_school_open = []
-
-	# call the simulations and save
-	matrix_infected, matrix_death, matrix_recovery, matrix_vaccination = main_algorithm_fast(n_simulations, n_days, n_nodes, n_initial_infected, array_network_parameters, array_weights)
-	for i in range(n_simulations):
-		matrix_infected[i,0] += n_initial_infected
-	matrix_infected, matrix_death, matrix_recovery, matrix_vaccination, n_non_spread = filter_non_spread(matrix_infected, matrix_death, matrix_recovery, matrix_vaccination)
-	print(n_non_spread)
-	n_simulations -= n_non_spread
-	plot_results("reults_witout_non_spread.pdf",matrix_infected, matrix_death, matrix_recovery, matrix_vaccination)
-	save_results(matrix_infected, matrix_death, matrix_recovery, matrix_vaccination)
-
-
-
-
-
-
-
-
-
