+++ non_spread_array.csv
@@ -1,0 +1,5 @@
+,0
+0,2.0
+1,0.0
+2,3.0
+3,2.0
