+++ parameters_age.csv
@@ -1,0 +1,21 @@
+﻿Age,population_UK,employment_rate,proportion_of_key_workers,infection_fatality_ratio
+0 - 4,3857263.00,0,0,0.0002
+5 - 9,4149852.00,0,0,0.0002
+10 - 14,3953866.00,0,0,0.0002
+15 - 19,3656968.00,0.252,0.019,0.0002
+20 - 24,4153080.00,0.611,0.073,0.0002
+25 - 29,4514249.00,0.847,0.115,0.0002
+30 - 34,4497132.00,0.847,0.12,0.0002
+35 - 39,4395667.00,0.849,0.118,0.0002
+40 - 44,4019539.00,0.849,0.113,0.0002
+45 - 49,4402122.00,0.849,0.121,0.0045
+50 - 54,4661015.00,0.711,0.122,0.0045
+55 - 59,4405908.00,0.711,0.104,0.0045
+60 - 64,3755185.00,0.711,0.063,0.0045
+65 - 69,3368199.00,0.106,0.02,0.0257
+70 - 74,3318867.00,0.106,0.012,0.0257
+75 - 79,2325296.00,0.106,0,0.156
+80 - 84,1715328.00,0.106,0,0.156
+85 - 89,1042090.00,0.106,0,0.156
+90+ (most of data don't specify age),605181.00,0.106,0,0.156
+,,,,
