+++ parameters_day.csv
@@ -1,0 +1,23 @@
+﻿Day_of_infection,Incubation_Period,Probability_of_Death,Probability_of_infecting,Probability_of_Recovery,Probability_of_Hospitilisation
+1,0.00625,0.001266242,0,14,0.001477283
+2,0.075,0.002864461,0.15,14,0.003798662
+3,0.15,0.005971954,0.375,14,0.00874063
+4,0.18,0.011474594,0.675,14,0.017996989
+5,0.1625,0.020319154,0.75,14,0.033159046
+6,0.125,0.033160516,0.6,14,0.054670025
+7,0.09,0.049875164,0.5,14,0.080656908
+8,0.07,0.069134493,0.4,14,0.106482669
+9,0.05,0.088318689,0.35,14,0.125794409
+10,0.035,0.10398192,0.25,14,0.13298076
+11,0.025,0.112826326,0.15,14,0.125794409
+12,0.013,0.112826326,0.125,14,0.106482669
+13,0.0125,0.10398192,0.1,14,0.080656908
+14,0.012,0.088318689,0.08,14,0.054670025
+15,0.007,0.069134493,0.06,14,0.033159046
+16,0.00625,0.049875164,0.06,14,0.017996989
+17,0.006,0.033160516,0.05,14,0.00874063
+18,0.006,0.020319154,0,14,0.003798662
+19,0.005,0.011474594,0,14,0.001477283
+20,0,0.005971954,0,14,0
+21,0,0.002864461,0,14,0
+22,0,0.001266242,0,14,0
