+++ parameters_matrix.csv
@@ -1,0 +1,5 @@
+,0,1,2,3,4
+0,2,2,2,2,2
+1,2,2,2,1,1
+2,2,2,1,1,1
+3,1,1,1,1,1
