+++ recovery_parameters_results.csv
@@ -1,0 +1,5 @@
+,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89
+0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.125,0.0,1.875,0.5,0.375,1.125,1.25,0.875,1.125,1.875,2.625,2.5,2.75,3.0,2.25,4.5,3.875,4.25,5.375,6.875,6.75,8.25,12.875,10.0,12.0,12.75,11.25,16.0,17.875,17.75,20.0,21.75,16.0,19.875,18.375,20.75,17.75,20.375,20.5,19.875,18.625,21.625,18.75,18.625,18.875,16.875,16.75,16.125,18.25,16.0,15.5,14.5,13.875,14.375,14.375,13.5,13.0,10.25,11.625,11.75,10.75,9.25,10.125,7.25,10.25,8.375,8.375,5.75,6.125,6.0,8.375,6.375,4.25,6.25,5.375,6.375,4.75,5.75
+1,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.1,2.0,0.1,0.5,0.8,0.6,0.9,1.3,0.6,2.4,2.2,3.0,2.0,2.1,2.5,3.8,3.8,4.7,4.0,4.8,5.4,5.1,5.7,7.6,6.2,7.1,8.0,7.3,8.9,10.3,7.1,8.8,8.8,7.3,10.2,11.1,9.3,10.7,10.6,10.0,11.8,11.4,9.5,10.0,10.5,11.2,12.7,10.5,12.8,10.9,10.3,9.9,9.9,9.8,9.3,8.9,10.6,9.8,9.0,9.0,10.5,8.7,7.2,11.3,9.0,7.6,9.0,7.1,9.4,7.4,7.8,7.9,6.8,7.6,7.1,6.6,7.3
+2,0.0,0.0,0.0,0.0,0.0,0.14285714285714285,0.0,0.0,0.14285714285714285,0.0,0.0,0.0,0.0,0.0,2.142857142857143,0.7142857142857143,0.5714285714285714,0.8571428571428571,1.1428571428571428,1.2857142857142858,1.8571428571428572,2.0,0.7142857142857143,2.2857142857142856,2.5714285714285716,2.5714285714285716,3.857142857142857,4.857142857142857,5.714285714285714,6.714285714285714,3.857142857142857,4.857142857142857,5.0,4.857142857142857,5.857142857142857,5.857142857142857,5.857142857142857,4.857142857142857,6.285714285714286,5.571428571428571,5.285714285714286,5.714285714285714,5.285714285714286,7.285714285714286,9.285714285714286,8.0,8.714285714285714,7.428571428571429,7.0,9.857142857142858,10.0,11.571428571428571,9.142857142857142,6.857142857142857,8.571428571428571,8.571428571428571,10.714285714285714,11.428571428571429,11.142857142857142,8.714285714285714,10.571428571428571,10.285714285714286,11.142857142857142,11.142857142857142,9.142857142857142,9.714285714285714,8.857142857142858,11.0,9.285714285714286,8.714285714285714,8.428571428571429,9.142857142857142,8.857142857142858,8.714285714285714,9.428571428571429,10.285714285714286,9.571428571428571,10.0,8.428571428571429,8.428571428571429,9.142857142857142,8.857142857142858,7.571428571428571,8.857142857142858,7.0,6.714285714285714,5.428571428571429,6.142857142857143,7.285714285714286,6.714285714285714
+3,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,1.75,0.375,0.5,1.0,1.0,0.5,1.125,2.125,1.5,1.5,1.25,1.75,2.625,3.0,2.625,4.0,3.625,3.875,3.5,3.75,3.625,3.5,3.75,4.375,4.875,5.25,3.875,3.625,3.5,4.375,4.125,4.125,5.25,5.0,3.625,3.875,5.125,3.375,3.375,3.75,4.875,5.375,4.5,3.25,4.375,6.0,3.875,3.25,4.5,5.0,3.875,5.25,3.375,4.5,3.25,1.625,3.375,3.875,3.625,5.625,5.125,3.25,4.125,3.375,3.75,3.125,4.25,3.125,3.875,4.25,4.625,5.25,3.875,3.125,3.75,4.0
