+++ recovery_results.csv
@@ -1,0 +1,6 @@
+,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89
+0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,2.0,1.0,0.0,2.0,5.0,3.0,2.0,4.0,6.0,2.0,6.0,6.0,3.0,10.0,13.0,8.0,5.0,12.0,14.0,14.0,21.0,26.0,21.0,40.0,38.0,25.0,34.0,34.0,33.0,26.0,40.0,31.0,32.0,26.0,29.0,23.0,21.0,24.0,30.0,20.0,30.0,15.0,31.0,21.0,20.0,18.0,14.0,12.0,16.0,6.0,4.0,9.0,2.0,8.0,4.0,4.0,9.0,6.0,2.0,7.0,8.0,5.0,2.0,8.0,4.0,1.0,2.0,4.0,3.0,3.0,1.0,1.0,4.0,0.0,5.0,1.0
+1,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,1.0,0.0,3.0,0.0,0.0,0.0,0.0,2.0,3.0,4.0,3.0,1.0,1.0,4.0,0.0,3.0,5.0,4.0,2.0,2.0,3.0,5.0,10.0,12.0,6.0,10.0,13.0,18.0,15.0,26.0,23.0,16.0,21.0,27.0,23.0,23.0,26.0,28.0,33.0,27.0,26.0,27.0,30.0,27.0,30.0,32.0,37.0,22.0,28.0,24.0,25.0,20.0,17.0,7.0,19.0,11.0,14.0,8.0,12.0,13.0,13.0,18.0,17.0,8.0,14.0,6.0,5.0,8.0,6.0,3.0,6.0,3.0,5.0,3.0,2.0,5.0,0.0,5.0
+2,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,5.0,0.0,1.0,0.0,2.0,1.0,2.0,4.0,2.0,1.0,0.0,0.0,1.0,1.0,5.0,2.0,3.0,3.0,2.0,2.0,6.0,5.0,9.0,7.0,4.0,2.0,8.0,11.0,14.0,17.0,26.0,18.0,21.0,13.0,23.0,19.0,26.0,31.0,26.0,30.0,22.0,24.0,27.0,27.0,34.0,33.0,32.0,21.0,24.0,24.0,16.0,19.0,19.0,11.0,9.0,6.0,13.0,7.0,9.0,5.0,8.0,15.0,4.0,6.0,4.0,12.0,4.0,6.0,3.0,9.0,4.0,3.0,4.0,10.0,3.0,2.0
+3,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,6.0,0.0,0.0,0.0,1.0,2.0,0.0,3.0,2.0,1.0,1.0,1.0,2.0,0.0,2.0,0.0,0.0,3.0,6.0,4.0,6.0,11.0,11.0,10.0,9.0,18.0,23.0,13.0,10.0,12.0,22.0,35.0,33.0,21.0,22.0,29.0,30.0,32.0,26.0,31.0,21.0,22.0,21.0,19.0,22.0,19.0,24.0,22.0,21.0,16.0,15.0,14.0,15.0,20.0,12.0,6.0,9.0,9.0,9.0,12.0,4.0,3.0,5.0,5.0,7.0,5.0,4.0,3.0,3.0,3.0,2.0,2.0,3.0,4.0,5.0,3.0
+4,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,4.0,0.0,3.0,3.0,1.0,2.0,5.0,0.0,2.0,2.0,5.0,6.0,2.0,3.0,2.0,2.0,14.0,9.0,6.0,15.0,13.0,18.0,15.0,18.0,22.0,26.0,25.0,28.0,28.0,33.0,31.0,38.0,33.0,41.0,28.0,22.0,28.0,17.0,27.0,18.0,21.0,22.0,17.0,20.0,13.0,14.0,11.0,7.0,4.0,13.0,11.0,7.0,7.0,7.0,6.0,6.0,5.0,2.0,1.0,3.0,2.0,1.0,1.0,2.0,1.0,4.0,0.0,0.0,1.0,2.0,0.0,0.0,0.0,1.0,1.0,0.0
